Multi-executor benchmark variant + nsys finalize fix + NITROS sub resolution
Drives the gantt visualization to its first full cross-executor session
(stereo_image_proc with 7 separate ComposableNodeContainers) and fixes the
attribution gaps surfaced by it.

Workload: one-container-per-node variant
- docker/custom_scripts/isaac_ros_disparity_graph_custom.py:
  rewrites the stock disparity_graph to wrap every ComposableNode in its
  own ComposableNodeContainer (= separate process + MultiThreadedExecutor).
  Exposes cross-process comm edges and per-executor scheduling.
- docker/fish-r2b-stereo_image_proc-custom.Dockerfile:
  thin overlay on the existing fish-r2b-stereo_image_proc image that drops
  the new launch script into both src and install/share trees so launch_test
  finds it without colcon rebuild.

nsys finalize lifecycle fix (multi-container case)
- python/fish/launch_wrap.py: force-override ExecuteProcess sigterm_timeout
  unconditionally (was conditional on kwargs absence). ros2_benchmark's
  with-nsys helper sets sigterm_timeout='1000' but downstream still resolved
  to 5s, causing nsys main to be SIGTERM'd 5s after SIGINT — mid-finalize —
  and the slow finalizers (DisparityNode, PointCloudNode, PlaybackNode)
  segfaulted (exit 139) before writing their .nsys-rep / .sqlite.
- scripts/install_fish.sh: replace the old "events stable for 2s" drain
  loop with a robust file-count drain
    * counts .nsys-rep files in session_dir/nsys/ — binary-agnostic
    * minimum wait FISH_NSYS_MIN_WAIT (default raised from 15 → much higher
      so finalize subprocesses (QdstrmImporter etc.) survive past launch teardown
    * pgrep matches nsys/QdstrmImporter under nsight-systems install only —
      not session-dir-in-cmdline, which previously caught LTTng and deadlocked
    * fixes `|| echo 0` double-fire bug that produced "0\n0" counts
- scripts/init_fish_pg.sql: widen all gpu_*.stream_id columns INT → BIGINT
  (CUDA emits 0xFFFFFFFF = uint32 sentinel which overflows signed int).
  Existing PG cluster: ALTER TABLE applied separately at runtime.
Result: 6/6 .nsys-rep + 6/6 .sqlite files captured for the stereo session.

NITROS-aware sub_handle → callback resolution
- postprocess/model_improved_pg.py: drop _keep_first for sub_init →
  introduce _resolve_sub_handle_to_cb() that:
    1. groups ALL (sub_handle → [sub_obj…]) and (sub_obj → [callback…])
    2. for handles with multiple candidate callbacks, picks the one whose
       cb_addr actually fires in the trace (callback_start events)
    3. logs WARN with category counters at the end
  Also adds warn() / _warn_summary() helpers used by other passes.
  Closes a real attribution gap: 16 NITROS-negotiated subscriptions whose
  callbacks fired ~14.5k times were silently missed by the old keep-first
  heuristic; now bound correctly.

Gantt API + viz upgrades
- postprocess/fish_viz_server.py:
  - /api/gantt SQL adds a tracepoint-based fallback for cb_addrs the model
    misses (defence in depth — model fix should drive the count to zero)
  - pid → node mapping now derives from ros2_trace.rcl_node_init since
    graph_nodes.N rows have pid=NULL
  - %% escapes in LIKE patterns (psycopg2 placeholder collision)
  - WARN to stderr on every fallback hit + (unknown) span count
- postprocess/gantt.html:
  - vertical scroll wrapper: when vtid count overflows the viewport,
    minimum 18px lane height kicks in and the page scrolls
  - x-axis moved to TOP (axisTop) so it stays visible while scrolling lanes
  - lanes sorted by (pid, vtid) instead of just vtid
  - per-pid color coding via d3.schemeTableau10/Paired
  - tinted left-margin band + small swatch per lane label
  - lane-label hover shows {vtid, pid, process_name}
  - thin horizontal dividers at pid boundaries (LaTeX-cline style)

Verified end-to-end on stereo_custom_full_20260622_184919:
  64K ros2 events, 897 graph vertices, 1354 edges (53 L0 cross-executor),
  20454 Tempo spans / 14 traces, 0 unresolved span attributions.

+++ scripts/init_fish_pg.sql
@@ -167,7 +167,7 @@ CREATE TABLE IF NOT EXISTS cuda_context (
     device_id         INT,
     parent_context_id INT,
     is_green_context  BOOLEAN,
-    null_stream_id    INT,
+    null_stream_id BIGINT,
     num_multiprocessors INT,
     PRIMARY KEY (session_id, context_id)
 );
@@ -184,7 +184,7 @@ CREATE TABLE IF NOT EXISTS cuda_device (
 
 CREATE TABLE IF NOT EXISTS cuda_streams (
     session_id  TEXT NOT NULL REFERENCES sessions(session_id) ON DELETE CASCADE,
-    stream_id   INT  NOT NULL,
+    stream_id   BIGINT NOT NULL,
     hw_id       INT,
     vm_id       INT,
     process_id  INT,
@@ -304,7 +304,7 @@ CREATE TABLE IF NOT EXISTS gpu_kernels (
     correlation_id      BIGINT,
     context_id          INT,
     device_id           INT,
-    stream_id           INT,
+    stream_id BIGINT,
     global_pid          BIGINT,
     grid_x              INT, grid_y INT, grid_z INT,
     block_x             INT, block_y INT, block_z INT,
@@ -326,7 +326,7 @@ CREATE TABLE IF NOT EXISTS gpu_memcpy (
     ts TIMESTAMPTZ GENERATED ALWAYS AS (to_timestamp(ts_ns::double precision / 1000000000)) STORED,
     session_id TEXT NOT NULL REFERENCES sessions(session_id) ON DELETE CASCADE,
     duration_ns BIGINT, bytes BIGINT, copy_kind INT,
-    correlation_id BIGINT, context_id INT, device_id INT, stream_id INT,
+    correlation_id BIGINT, context_id INT, device_id INT, stream_id BIGINT,
     global_pid BIGINT, container TEXT, source TEXT,
     PRIMARY KEY (session_id, ts_ns, id)
 );
@@ -337,7 +337,7 @@ CREATE TABLE IF NOT EXISTS gpu_memset (
     ts TIMESTAMPTZ GENERATED ALWAYS AS (to_timestamp(ts_ns::double precision / 1000000000)) STORED,
     session_id TEXT NOT NULL REFERENCES sessions(session_id) ON DELETE CASCADE,
     duration_ns BIGINT, bytes BIGINT, value BIGINT, mem_kind INT,
-    correlation_id BIGINT, context_id INT, device_id INT, stream_id INT,
+    correlation_id BIGINT, context_id INT, device_id INT, stream_id BIGINT,
     global_pid BIGINT, container TEXT, source TEXT,
     PRIMARY KEY (session_id, ts_ns, id)
 );
@@ -349,7 +349,7 @@ CREATE TABLE IF NOT EXISTS gpu_sync (
     session_id TEXT NOT NULL REFERENCES sessions(session_id) ON DELETE CASCADE,
     duration_ns BIGINT,
     sync_type INT, event_id BIGINT, event_sync_id BIGINT,
-    correlation_id BIGINT, context_id INT, device_id INT, stream_id INT,
+    correlation_id BIGINT, context_id INT, device_id INT, stream_id BIGINT,
     global_pid BIGINT, container TEXT, source TEXT,
     PRIMARY KEY (session_id, ts_ns, id)
 );
